+++ ChestsAnywhere/docs/release-notes.md
@@ -1,6 +1,9 @@
 ﻿[← back to readme](README.md)
 
 # Release notes
+## Upcoming release
+* When a custom location has a very long name, it's now truncated when used as a default chest category.
+
 ## 1.25.1
 Released 04 November 2024 for SMAPI 4.1.2 or later.